+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/config/publishing/_internal/configureMavenPublishing.kt
@@ -195,7 +195,7 @@ private fun HubdleConfigurableExtension.configureSigningForPublishing() =
             val shouldSign = getPropertyOrNull(HubdleProperty.Publishing.sign)?.toBoolean() ?: false
 
             val hasTaskCondition = (hasPublishTask && !hasPublishToMavenLocalTasks)
-            val hasSemverCondition = project.isNotSnapshot && project.isSemver
+            val hasSemverCondition = project.isNotSnapshot.get() && project.isSemver
 
             isRequired = (hasTaskCondition && hasSemverCondition) || shouldSign