+++ hubdle-gradle-plugin/test/kotlin/com/javiersc/hubdle/_utils/resourceFile.kt
@@ -1,0 +1,6 @@
+package com.javiersc.hubdle._utils
+
+import java.io.File
+
+internal fun resourceFile(resource: String): File =
+    File(Thread.currentThread().contextClassLoader?.getResource(resource)?.toURI()!!)