+++ hubdle-gradle-plugin/test/kotlin/com/javiersc/hubdle/config/BinaryCompatibilityStateTest.kt
@@ -1,13 +1,13 @@
 package com.javiersc.hubdle.config
 
-import com.javiersc.gradle.testkit.test.extensions.GradleTest
+import com.javiersc.gradle.testkit.test.extensions.GradleTestKitTest
 import io.kotest.matchers.file.shouldBeADirectory
 import io.kotest.matchers.file.shouldBeAFile
 import io.kotest.matchers.string.shouldContain
 import java.io.File
 import kotlin.test.Test
 
-internal class BinaryCompatibilityStateTest : GradleTest() {
+internal class BinaryCompatibilityStateTest : GradleTestKitTest() {
 
     private val basePath = "config/binary-compatibility-validator"