+++ hubdle-gradle-plugin/test/kotlin/com/javiersc/hubdle/kotlin/multiplatform/KotlinMultiplatformTest.kt
@@ -1,6 +1,6 @@
 package com.javiersc.hubdle.kotlin.multiplatform
 
-import com.javiersc.gradle.testkit.test.extensions.GradleTest
+import com.javiersc.gradle.testkit.test.extensions.GradleTestKitTest
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.file.shouldBeADirectory
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable
 import org.junit.jupiter.api.condition.EnabledOnOs
 import org.junit.jupiter.api.condition.OS
 
-internal class KotlinMultiplatformTest : GradleTest() {
+internal class KotlinMultiplatformTest : GradleTestKitTest() {
 
     private val basePath = "kotlin/multiplatform"